Move the sendEmail function into its own module
